+++ doc/version_history.rst
@@ -6,6 +6,26 @@
 Version History
 ##################
 
+.. _lsst.ts.hexgui-0.3.0:
+
+-------------
+0.3.0
+-------------
+
+* Depend on the **ts_hexrotcomm**.
+* Copy the **structs.py** from the **ts_mthexapod**.
+* Add the limits to **contants.py**.
+* Update the enum value to be consistent with the **ts_hexapod_controller**.
+* Remove the **config.py**.
+* Fix the input pin in **TabDriveStatus**.
+* Use the mirco degree in **TabPosition**.
+* Support the TCP/IP communication with the controller.
+* Update the range of parameters in **ControlPanel**.
+* Update the **MainWindow** to connect/disconnect the controller.
+* Extract the **TabConfig** from **TabTelemetry**.
+* Update the **TabTelemetry** to show the current data.
+* Remove the log button in **ControlPanel**.
+
 .. _lsst.ts.hexgui-0.2.0:
 
 -------------